🐛 Revert changes, fix int64_t

+++ paddle/phi/kernels/legacy/gpu/fast_rmsnorm_kernel.cu
@@ -38,7 +38,7 @@ void RMSLnFwdKernel(const Context& dev_ctx,
   auto sizes = x.dims();
   PD_CHECK(sizes.size() >= 2);
 
-  const int cols = sizes[sizes.size() - 1];
+  const int64_t cols = sizes[sizes.size() - 1];
   const int64_t rows = x.numel() / cols;
   auto hidden_size = scale.numel();
